+++ src/js/support.js
@@ -10,7 +10,7 @@ const inputTypes = {};
 let mobile = false;
 
 // test input types
-[ 'date', 'datetime', 'time', 'month' ].forEach( inputType => {
+[ 'date', 'datetime-local', 'time', 'month' ].forEach( inputType => {
     const input = document.createElement( 'input' );
     input.setAttribute( 'type', inputType );
     inputTypes[ inputType ] = input.type !== 'text';